Move data tests to /data
