+++ test/data/test_dataset.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-"""Unit tests for cartoframes.context"""
+"""Unit tests for cartoframes.data.Dataset"""
 import unittest
 import os
 import sys
@@ -10,7 +10,7 @@ import pandas as pd
 
 from carto.exceptions import CartoException
 
-from cartoframes.auth import Context, Credentials
+from cartoframes.auth import Credentials
 from cartoframes.data import Dataset
 from cartoframes.data.utils import setting_value_exception
 from cartoframes.columns import normalize_name
@@ -20,10 +20,9 @@ from cartoframes.data.registry.dataframe_dataset import DataFrameDataset, _rows
 from cartoframes.data.registry.table_dataset import TableDataset
 from cartoframes.data.registry.query_dataset import QueryDataset
 
-from .mocks.dataset_mock import DatasetMock
-from .mocks.context_mock import ContextMock
+from ..mocks.dataset_mock import DatasetMock
 
-from .utils import _UserUrlLoader
+from ..utils import _UserUrlLoader
 
 try:
     import geopandas
